filesystem overhaul for db troubleshooting
