+++ .gitignore
@@ -1,7 +1,7 @@
 # Special
-api/config.py
-api/constants.py
-api/migrations/
+config.py
+constants.py
+migrations/
 # Created by https://www.toptal.com/developers/gitignore/api/python,react,flask
 # Edit at https://www.toptal.com/developers/gitignore?templates=python,react,flask
