+++ api/routes/login.py
@@ -2,9 +2,8 @@ from flask import request, abort, jsonify, Response, redirect
 from sqlalchemy import text
 import sys
 sys.path.append('../..')
-from api import db
-from __init__ import routes
-import models
+from api import db, models
+from api.routes import routes
 from cerberus import Validator
 from flask_login import login_user
