+++ main.py
@@ -1,0 +1,6 @@
+from api import app, db
+from api.models import User, Device
+
+@app.shell_context_processor
+def make_shell_context():
+    return {'db': db, 'User': User, 'Device': Device}
